[Decomp][Player] Cohesion split: extract talent system to PlayerTalent.cpp
